+++ spec/spec_helper.rb
@@ -1,6 +1,18 @@
 # frozen_string_literal: true
 
-require "bundler/setup"
+ENV["BUNDLE_GEMFILE"] ||= File.expand_path("../Gemfile", __dir__)
+
+require "bundler/setup" if File.exist?(ENV["BUNDLE_GEMFILE"])
+
+begin
+  require "simplecov"
+  SimpleCov.start do
+    add_filter ["/spec/"]
+  end
+rescue LoadError
+end
+
+Bundler.require(:default, :test)
 
 require "climate_control"
 require "nokogiri"